requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-simple-serializer==2.0.7
djangorestframework==3.9.2
future==0.17.1
mysqlclient==1.4.2.post1
Pygments==2.3.1
pytz==2019.1
sqlparse==0.3.0
